+++ Arcane Waters - Alpha/Assets/Scripts/Network/DeployScheduleInfo.cs
@@ -4,12 +4,12 @@
         /// <summary>
         /// The scheduled date represented as ticks.
         /// </summary>
-        public int DateAsTicks;
+        public long DateAsTicks;
         /// <summary>
         /// The Build's Version Number.
         /// </summary>
         public int Build;
-        public DeployScheduleInfo(int dateAsTicks, int build)
+        public DeployScheduleInfo(long dateAsTicks, int build)
         {
             DateAsTicks = dateAsTicks;
             Build = build;
